requirements.txt
pyserial~=3.5
PySimpleGUI~=4.49.0
setuptools~=57.0.0